added borrow tool functionality
