+++ server.js
@@ -7,37 +7,6 @@ const helpers = require('./utils/helpers');
 
 require('dotenv').config();
 
-const nodemailer = require('nodemailer');
-const log = console.log;
-
-
-
-// Step 1
-let transporter = nodemailer.createTransport({
-  service: 'gmail',
-  auth: {
-      user: process.env.EMAIL,// || //'abc@gmail.com', // TODO: your gmail account
-      pass: process.env.PASSWORD //|| '1234' // TODO: your gmail password
-  }
-});
-
-// Step 2
-let mailOptions = {
-  from: 'toolin.around21@gmail.com', // TODO: email sender
-  to: 'toolin.around21@gmail.com', // TODO: email receiver
-  subject: 'Nodemailer - Test',
-  text: 'Wooohooo'
-};
-
-// Step 3
-transporter.sendMail(mailOptions, (err, data) => {
-  if (err) {
-      return log('Error occurs', err);
-  }
-  return log('Email sent!!!');
-});
-
-
 const sequelize = require('./config/connection');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
